requirements.txt
web3>=6.15.0
httpx>=0.26.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
